Aktualizacja plików związanych z ankietami
