+++ workspaces/malachapp/lib/pages/home.dart
@@ -13,9 +13,9 @@ import 'package:malachapp/pages/creator.dart';
 import 'package:malachapp/pages/event_page.dart';
 import 'package:malachapp/pages/home_home.dart';
 import 'package:malachapp/pages/message_broadcast_page.dart';
-import 'package:malachapp/pages/poll_list_design.dart';
+import 'package:malachapp/pages/Poll/poll_list_design.dart';
 // import 'package:malachapp/pages/home_home.dart';
-import 'package:malachapp/pages/poll_page.dart';
+import 'package:malachapp/pages/Poll/poll_page.dart';
 import 'package:malachapp/services/fb_storage_loader.dart';
 import 'package:malachapp/services/notification_service.dart';
 // import 'package:malachapp/services/notification_service.dart';
